Properly encapsulate RequestBuilder (nested class, etc.)
No public constructor for either Request or RequestBuilder, forces all
uses through the validating public API.

Also added some more (failing) tests
